Fixed Ride packet write.

+++ L2J_Mobius_05.0_Salvation/java/org/l2jmobius/gameserver/network/serverpackets/Ride.java
@@ -42,7 +42,7 @@ public class Ride extends ServerPacket
 	{
 		ServerPackets.RIDE.writeId(this);
 		writeInt(_objectId);
-		writeByte(_mounted);
+		writeInt(_mounted);
 		writeInt(_rideType);
 		writeInt(_rideNpcId);
 		writeInt(_loc.getX());
